+++ test/hello-test.ts
@@ -1,6 +1,5 @@
 import { url } from "../src/setup-server";
 import { assert } from "chai";
-import { describe, it } from "mocha";
 import axios from "axios";
 
 describe("Hello Query Test", function () {
