+++ test/test-queries.ts
@@ -1,11 +1,5 @@
-import { LoginInput, UserInput } from "../../src/graphql/schema";
+import { LoginInput, UserInput } from "../src/graphql/schema";
 import bcrypt from "bcrypt";
-
-export const hashPassword = async (password: string) => {
-  const saltRounds = 10;
-  const hashedPassword = await bcrypt.hash(password, saltRounds);
-  return hashedPassword;
-};
 
 export function loginUserMutation(data: LoginInput) {
   return `#graphql
