+++ test/test.ts
@@ -1,4 +1,4 @@
-import { describe, before, after } from "mocha"; // Mocha imports
+import { describe, before, after } from "mocha";
 import { setup } from "../src/setup";
 import { server } from "../src/setup-server";
 import { prisma } from "../src/setup-db";
